refactor: consolidate utils directory
