+++ __tests__/bff/cms-content.test.js
@@ -1,6 +1,6 @@
 import { vi, describe, it, expect, beforeEach, afterEach } from 'vitest';
 import { getCMSContent } from '../../bff/cms-content/index.js';
-import * as fetchDataModule from '../../bff/utils/fetchData.js';
+import * as fetchDataModule from '../../src/utils/fetchData.js';
 
 // Mock applicationinsights
 const mockTrackTrace = vi.fn();
